requirements.txt
fastapi==0.129.2
uvicorn==0.41.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.1
alembic==1.18.4